% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{initial_records}
\alias{initial_records}
\title{Prepare data for use within the app}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
initial_records
}
\description{
Prepare data for use within the app
}
\keyword{datasets}
